Do not exit the application in case the settings cannot be read
Previously, issues such as deserialization problems or the absence of
the settings file would trigger an exception, leading to a forcefully
application termination.

This patch addresses this issue by removing the timer that triggers the
application exit, since the calling code is fully exception-aware. This
also allows application resources to perform any cleanup or finalization
necessary once the user closes the application.