Introduce library to host common code shared by FroniusMonitor and FroniusPhone
Move IDataCollectionService, DataCollectionService, HomeAutomationSystem,
SolarDataEventArgs, SmartMeterCalibrationHistoryItem and SettingsBase into
a shared libray that hosts common code for both applications. All code is
really tied to the application that is build on top of the Fronius library,
nothing that can be reused per se by a different home automation system or
a different setup of devices.

# Conflicts:
#	FroniusMonitor.sln